+++ src/index.ts
@@ -1,6 +1,16 @@
 import type { Config } from 'prettier';
 
-export default {
+export const twig = {
+  twigMultiTags: ['html_element,end_html_element', 'with,endwith', 'apply,endapply'],
+  plugins: [import.meta.resolve('@zackad/prettier-plugin-twig')],
+} satisfies Config;
+
+export const liquid = {
+  singleQuote: false,
+  plugins: [import.meta.resolve('@shopify/prettier-plugin-liquid')],
+} satisfies Config;
+
+export const config = {
   bracketSpacing: true,
   printWidth: 100,
   semi: true,
@@ -14,18 +24,13 @@ export default {
   overrides: [
     {
       files: '**/*.twig',
-      options: {
-        twigPrintWidth: 100,
-        twigMultiTags: ['html_element,end_html_element', 'with,endwith', 'apply,endapply'],
-        plugins: [import.meta.resolve('@afshinhaghighat/prettier-plugin-twig-melody')],
-      },
+      options: twig,
     },
     {
       files: '**/*.liquid',
-      options: {
-        singleQuote: false,
-        plugins: [import.meta.resolve('@shopify/prettier-plugin-liquid')],
-      },
+      options: liquid,
     },
   ],
 } satisfies Config;
+
+export default config;